Make character_encoding.cc compile on Mac and Linux. Review URL: http://codereview.chromium.org/21220
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@9513 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/character_encoding.cc
@@ -20,7 +20,7 @@
 namespace {
 
 // The maximum length of short list of recently user selected encodings is 3.
-const int kUserSelectedEncodingsMaxLength = 3;
+const size_t kUserSelectedEncodingsMaxLength = 3;
 
 typedef struct {
   int resource_id;
@@ -431,4 +431,3 @@ bool CharacterEncoding::UpdateRecentlySelectdEncoding(
   }
   return true;
 }
-
